refactor(civitai): split base-model helpers into basemodels_civitai
Addresses #4756 review: keeps client_civitai focused on the HTTP client
surface. Behavior unchanged.
